requirements.txt
BeautifulSoup==3.2.0
Django==1.2.3
Werkzeug==0.6.2
South==0.7.3
amqplib==0.6.1
anyjson==0.3
carrot==0.10.7
celery==2.1.4
distribute==0.6.14
django-ajaxerrors==1.1.0
django-annoying==0.7.5
django-celery==2.1.4
django-css==2.3.1
django-debug-toolbar==0.8.4
django-devserver==0.1.4
django-extensions==0.6
django-indexer==0.2.1
django-paging==0.2.2
django-picklefield==0.1.9
django-sentry==1.4.1
django-templatetag-sugar==0.1
flup==1.0.2
gargoyle==0.1.0
guppy==0.1.9
httplib2==0.6.0
importlib==1.0.2
mimeparse==0.1.3
oauth2==1.2.0
psycopg2==2.3.1
pyparsing==1.5.5
python-dateutil==1.5
scikits.learn==0.6
simplejson==2.1.2
sqlparse==0.1.2
virtualenv==1.5.1
virtualenvwrapper==2.5.3
wsgiref==0.1.2
yolk==0.4.1

